Breakout math functions into csp.math, tweak edge ufunc dispatch to use big dict
